Cleaned up bike list screen and made the style consistent with the rest of the app

+++ lib/screens/bike_details_screen.dart
@@ -201,7 +201,7 @@ class _BikeDetailsScreenState extends State<BikeDetailsScreen> {
 
   double _getRating() {
     return _bike.data.containsKey('${AppStrings.bikeRatingKey}')
-        ? _bike['${AppStrings.bikeRatingKey}']
+        ? _bike['${AppStrings.bikeRatingKey}'].toDouble()
         : 0.0;
   }
